Allow support for specifying database to test on.
This only applies to the Django based tests, because normal python tests don't need a database.

This also added the ability to include the site packages inside the virtualenv. This is because psycopg2 and other database style adapters are hard to install, and it is annoying having to require them and install them for running against that databse.